+++ src/main/java/yqloss/yqlossclientmixinkt/impl/MainConfig.kt
@@ -16,6 +16,22 @@ import kotlin.reflect.full.declaredMemberProperties
 import kotlin.reflect.full.isSubtypeOf
 import kotlin.reflect.full.starProjectedType
 
+const val INFO_DISCLAIMER =
+    "Use all Yqloss Client features at your own risk! It's not guaranteed that you wouldn't be banned using \"safe\" features!"
+const val LEGITIMACY_LEGIT = "The features below are safe to use."
+const val LEGITIMACY_SAFE_QOL =
+    "The features below change vanilla behavior but are not detectable by servers. They are basically safe to use."
+const val LEGITIMACY_SAFE_MACRO =
+    "The features below are considered as macros but are not likely to get you banned on most servers."
+const val LEGITIMACY_UNSAFE_MACRO = "THE FEATURES BELOW ARE CONSIDERED AS MACROS AND CAN GET YOU BANNED!"
+const val LEGITIMACY_SAFE_BLATANT =
+    "The features below change vanilla behavior but are not likely to get you banned on most servers."
+const val LEGITIMACY_SAFE_BLATANT_HYPIXEL =
+    "The features below change vanilla behavior but are not likely to get you banned on Hypixel."
+const val LEGITIMACY_SAFE_BLATANT_SKYBLOCK =
+    "The features below change vanilla behavior but are not likely to get you banned on Hypixel SkyBlock."
+const val LEGITIMACY_UNSAFE_BLATANT = "THE FEATURES BELOW CHANGE VANILLA BEHAVIOR AND CAN GET YOU BANNED!"
+
 private val optionsImplMap = mutableMapOf<KClass<*>, () -> YCModuleOptions>()
 
 private val LOGGER = ycLogger("Config")